Update test.yml to use specific node version (14.x) (2)
Add a specifier in test.yml to use node version 14 due to Open Handler issue when running jest tests on node 16 (callstack/react-native-testing-library#1007)

Thank you for contributing to PEAR!
Before we proceed, please fill-in the following template for your PR:

If you haven't already created a new issue to reference to, please kindly create a new issue in the given trello board.
https://trello.com/b/i6t7NLzF/pear-fyp-sprint-board

Which issue is this PR related to?
Is this a feature? or a bug? or a Document?
If this is a new feature,(1) please explain what feature have you created and how does it value-add to our platform. (2) Copy and paste a screen shot on how the UI looks in the comment after making a PR.
If this is a bug, please explain what was the root cause of the error.
If this is a document, please explain the purpose of it.
Please complete the following checklist:
[] I have read up on the naming convention for my PR from: Our Contribution page. (If you haven't please read it from README.md in root.)
[] The feature/bug I have introduced will not disrupt the architecture of the platform.
[] I have run es-lint on all files to ensure the formatting is done right.
[] I have only committed the files I have changed in this PR.